+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 **Next release**
 
-* ...
+* Option to clear restrictions from notification (if your Android version supports this)
 
 **Version 1.9.23 release candidate**